refactor(core): extract framebuffer copy to separate bg_copy functionality
[no changelog]
